requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch~=2.4.0
torchvision~=0.19.0
torchmetrics~=1.4.0

scikit-learn~=1.5.0
numpy>=1.26,<3
tqdm>=4.66,<5

transformers>=4.42
pillow>=9.5
regex>=2023.10